feat: ストリーミング送信機能とテスト基盤の実装
- streaming.rs: ESP-NOWストリーミング送信機能実装
- camera/config.rs: XIAO ESP32S3用カメラ設定
- camera/ov2640.rs: OV2640センサー制御
- camera/xiao_esp32s3.rs: XIAO ESP32S3ピン設定
- tests/: TDDベースのテスト基盤構築
- 警告整理: unused import/code削除、将来実装用コードに#[allow(dead_code)]追加
- コンパイルエラー0件、テスト実行可能状態を達成

Issue #12 関連

+++ devices/xiao_esp32s3_sense/src/core/data_service.rs
@@ -2,7 +2,7 @@ use esp_idf_svc::hal::delay::FreeRtos;
 use log::{error, info, warn};
 
 use crate::communication::esp_now::EspNowSender;
-use crate::core::config::AppConfig;
+use crate::config::AppConfig;
 use crate::hardware::camera::{CameraController, CamConfig};
 use crate::hardware::led::StatusLed;
 
@@ -126,7 +126,7 @@ impl DataService {
         match esp_now_sender.send_image_chunks(
             image_data,
             app_config.esp_now_chunk_size as usize,  // 設定からチャンクサイズを取得
-            app_config.esp_now_chunk_delay_ms,  // 設定からチャンク間遅延を取得
+            app_config.esp_now_chunk_delay_ms as u32,  // 設定からチャンク間遅延を取得
         ) {
             Ok(_) => {
                 info!("画像データの送信が完了しました");
